Refactor paginated table.
