+++ app/components/TokensTable.tsx
@@ -1,7 +1,8 @@
-import React, { useState, useMemo, ReactElement } from 'react'
+import React, { ReactElement } from 'react'
 import { TokenBalance } from '../types/TokenBalance';
 import { TokenMetaData } from '@/app/types/TokenMetaData';
 import Image from 'next/image';
+import PaginatedTable from './PaginatedTable';
 
 interface TokensTableProps {
   tokenBalances: TokenBalance[];
@@ -9,47 +10,6 @@ interface TokensTableProps {
 }
 
 const TokensTable: React.FC<TokensTableProps> = ({ tokenBalances, tokensCache }) => {
-  const [currentPage, setCurrentPage] = useState(1);
-  const pageSize = 10;
-
-  // Calculate pagination values
-  const totalPages = Math.ceil(tokenBalances.length / pageSize);
-  const startIndex = (currentPage - 1) * pageSize;
-  const endIndex = startIndex + pageSize;
-  const currentTokens = tokenBalances.slice(startIndex, endIndex);
-
-  // Calculate which page numbers to show (5 pages centered around current page)
-  const getPageNumbers = useMemo(() => {
-    const pages: number[] = [];
-    let startPage = Math.max(1, currentPage - 2);
-    let endPage = Math.min(totalPages, currentPage + 2);
-
-    // Adjust if we're near the beginning or end
-    if (endPage - startPage < 4) {
-      if (startPage === 1) {
-        endPage = Math.min(totalPages, startPage + 4);
-      } else if (endPage === totalPages) {
-        startPage = Math.max(1, endPage - 4);
-      }
-    }
-
-    for (let i = startPage; i <= endPage; i++) {
-      pages.push(i);
-    }
-    return pages;
-  }, [currentPage, totalPages]);
-
-  // Navigation functions
-  const goToPage = (page: number) => {
-    if (page >= 1 && page <= totalPages) {
-      setCurrentPage(page);
-    }
-  };
-
-  const goToFirst = () => goToPage(1);
-  const goToPrevious = () => goToPage(currentPage - 1);
-  const goToNext = () => goToPage(currentPage + 1);
-  const goToLast = () => goToPage(totalPages);
 
   // Helper function to format value from wei to ETH
   const formatValue = (value: number, decimals: number, precision = 8): string => {
@@ -96,85 +56,30 @@ const TokensTable: React.FC<TokensTableProps> = ({ tokenBalances, tokensCache })
     )
   }
 
+  const renderHeader = () => {
+    <thead>
+      <tr>
+        <th></th>
+        <th>Token</th>
+        <th className="text-right">Balance</th>
+      </tr>
+    </thead>
+  }
+
+  const renderBody = (currentTokens: TokenBalance[]) => (
+    <tbody>
+      {currentTokens.map((token, index) => getTokenRow(token, index))}
+    </tbody>
+  );
+
   return (
-    <div className="overflow-x-auto">
-      <table className="table table-zebra w-full">
-        <thead>
-          <tr>
-            <th></th>
-            <th>Token</th>
-            <th className="text-right">Balance</th>
-          </tr>
-        </thead>
-        <tbody>
-          {currentTokens.map((token, index) => getTokenRow(token, index))}
-        </tbody>
-      </table>
-
-      {/* Pagination Controls */}
-      {totalPages > 1 && (
-        <div className="flex justify-center items-center mt-4 space-x-1">
-          {/* First Button */}
-          <button
-            onClick={goToFirst}
-            disabled={currentPage === 1}
-            className={`btn btn-sm ${currentPage === 1 ? 'btn-disabled' : 'btn-outline'}`}
-          >
-            First
-          </button>
-
-          {/* Previous Button */}
-          <button
-            onClick={goToPrevious}
-            disabled={currentPage === 1}
-            className={`btn btn-sm ${currentPage === 1 ? 'btn-disabled' : 'btn-outline'}`}
-          >
-            Previous
-          </button>
-
-          {/* Page Numbers */}
-          {getPageNumbers.map((pageNum) => (
-            <button
-              key={pageNum}
-              onClick={() => goToPage(pageNum)}
-              disabled={pageNum === currentPage}
-              className={`btn btn-sm ${
-                pageNum === currentPage
-                  ? 'btn-primary font-bold cursor-default'
-                  : 'btn-outline'
-              }`}
-            >
-              {pageNum}
-            </button>
-          ))}
-
-          {/* Next Button */}
-          <button
-            onClick={goToNext}
-            disabled={currentPage === totalPages}
-            className={`btn btn-sm ${currentPage === totalPages ? 'btn-disabled' : 'btn-outline'}`}
-          >
-            Next
-          </button>
-
-          {/* Last Button */}
-          <button
-            onClick={goToLast}
-            disabled={currentPage === totalPages}
-            className={`btn btn-sm ${currentPage === totalPages ? 'btn-disabled' : 'btn-outline'}`}
-          >
-            Last
-          </button>
-        </div>
-      )}
-
-      {/* Page Info */}
-      {totalPages > 1 && (
-        <div className="text-center mt-2 text-sm text-gray-600">
-          Showing {startIndex + 1}-{Math.min(endIndex, tokenBalances.length)} of {tokenBalances.length} transactions
-        </div>
-      )}
-    </div>
+    <PaginatedTable
+      data={tokenBalances}
+      pageSize={10}
+      renderHeader={() => renderHeader()}
+      renderBody={(currentData) => renderBody(currentData)}
+      itemName="tokens"
+    />
   )
 }
